Organize integration tests into crawl and tiktok suites
