+++ tests/integration/crawl/test_auspost_integration.py
@@ -9,7 +9,7 @@ from app.main import app
 
 
 # Ensure project root on sys.path for imports like app.*
-ROOT = Path(__file__).resolve().parents[2]
+ROOT = Path(__file__).resolve().parents[3]
 if str(ROOT) not in sys.path:
     sys.path.insert(0, str(ROOT))
