+++ tests/integration/crawl/test_crawl_real_urls_global.py
@@ -1,14 +1,12 @@
 import pytest
 
-from tests.integration._real_url_test_utils import make_body, min_html_length
+from tests.integration.crawl._real_url_test_utils import make_body, min_html_length
 
 pytestmark = pytest.mark.integration
 
 
-def test_crawl_ups(client):
-    body = make_body(
-        "https://www.ups.com/track?tracknum=1ZXH95910305309465&loc=vi_VN&requester=QUIC/trackdetails"
-    )
+def test_crawl_17track(client):
+    body = make_body("https://t.17track.net/en#nums=1ZXH95910326694965")
     resp = client.post("/crawl", json=body)
     assert resp.status_code == 200
     data = resp.json()
@@ -18,10 +16,8 @@ def test_crawl_ups(client):
     assert len(html) >= min_html_length()
 
 
-def test_crawl_fedex(client):
-    body = make_body(
-        "https://www.fedex.com/fedextrack/?trknbr=883561067070&trkqual=2460902000~883561067070~FX"
-    )
+def test_crawl_dpex(client):
+    body = make_body("https://dpexonline.com/trace-and-track/index?id=226006280426")
     resp = client.post("/crawl", json=body)
     assert resp.status_code == 200
     data = resp.json()
@@ -31,10 +27,19 @@ def test_crawl_fedex(client):
     assert len(html) >= min_html_length()
 
 
-def test_crawl_usps(client):
-    body = make_body(
-        "https://tools.usps.com/go/TrackConfirmAction?tRef=fullpage&tLc=2&text28777=&tLabels=9200190381836321489085%2C&tABt=false"
-    )
+def test_crawl_parcelsapp(client):
+    body = make_body("https://parcelsapp.com/en/tracking/9200190381599918197427")
+    resp = client.post("/crawl", json=body)
+    assert resp.status_code == 200
+    data = resp.json()
+    assert data.get("status") == "success"
+    html = data.get("html") or ""
+    assert "<html" in html.lower()
+    assert len(html) >= min_html_length()
+
+
+def test_crawl_laposte(client):
+    body = make_body("https://www.laposte.fr/outils/suivre-vos-envois?code=LA866151484GB")
     resp = client.post("/crawl", json=body)
     assert resp.status_code == 200
     data = resp.json()
