+++ tests/integration/crawl/test_crawl_real_urls_usa.py
@@ -1,12 +1,14 @@
 import pytest
 
-from tests.integration._real_url_test_utils import make_body, min_html_length
+from tests.integration.crawl._real_url_test_utils import make_body, min_html_length
 
 pytestmark = pytest.mark.integration
 
 
-def test_crawl_17track(client):
-    body = make_body("https://t.17track.net/en#nums=1ZXH95910326694965")
+def test_crawl_ups(client):
+    body = make_body(
+        "https://www.ups.com/track?tracknum=1ZXH95910305309465&loc=vi_VN&requester=QUIC/trackdetails"
+    )
     resp = client.post("/crawl", json=body)
     assert resp.status_code == 200
     data = resp.json()
@@ -16,8 +18,10 @@ def test_crawl_17track(client):
     assert len(html) >= min_html_length()
 
 
-def test_crawl_dpex(client):
-    body = make_body("https://dpexonline.com/trace-and-track/index?id=226006280426")
+def test_crawl_fedex(client):
+    body = make_body(
+        "https://www.fedex.com/fedextrack/?trknbr=883561067070&trkqual=2460902000~883561067070~FX"
+    )
     resp = client.post("/crawl", json=body)
     assert resp.status_code == 200
     data = resp.json()
@@ -27,19 +31,10 @@ def test_crawl_dpex(client):
     assert len(html) >= min_html_length()
 
 
-def test_crawl_parcelsapp(client):
-    body = make_body("https://parcelsapp.com/en/tracking/9200190381599918197427")
-    resp = client.post("/crawl", json=body)
-    assert resp.status_code == 200
-    data = resp.json()
-    assert data.get("status") == "success"
-    html = data.get("html") or ""
-    assert "<html" in html.lower()
-    assert len(html) >= min_html_length()
-
-
-def test_crawl_laposte(client):
-    body = make_body("https://www.laposte.fr/outils/suivre-vos-envois?code=LA866151484GB")
+def test_crawl_usps(client):
+    body = make_body(
+        "https://tools.usps.com/go/TrackConfirmAction?tRef=fullpage&tLc=2&text28777=&tLabels=9200190381836321489085%2C&tABt=false"
+    )
     resp = client.post("/crawl", json=body)
     assert resp.status_code == 200
     data = resp.json()
