+++ tests/integration/crawl/test_generic_crawl_integration.py
@@ -8,7 +8,7 @@ from fastapi.testclient import TestClient
 from app.main import app
 
 # Ensure project root on path
-ROOT = Path(__file__).resolve().parents[2]
+ROOT = Path(__file__).resolve().parents[3]
 if str(ROOT) not in sys.path:
     sys.path.insert(0, str(ROOT))
 # Integration tests use the real endpoint and fetcher
